[LLDB] Fix loading modules outside container first that causes the failure of stack unwinding and step-in command
With platform being "remote-hos inner", lldb may fail to resolve
modules properly because at first lldb tries to get shared libraries
residing outside the container. If a library with the same path but
different content is pulled, DWARF-based unwinding fails, and it
falls back into the unreliable frame-pointer-based unwinding (it is
not dependable because the frame pointer resiter may be used as a
general purpose register with the compiler flag -fomit-frame-pointer).
The issue can be fixed by retrieving shared libraries inside the
container instead.

Issue: I6EQG3
Test: build lldb and debug on devices
Signed-off-by: hongbinj <jinhongbin2@huawei.com>