[lldb][OHOS] Support remote HDC server
Adds support for connecting to an HDC server on a remote machine. The address
in the platform connection command is used as that remote machine address,
meaning that device id can now be only set with HDC_UTID environment variable.
The remote HDC server should be listening not on loopback to accept the lldb
connection, and the HDC forwarded ports should be accessible.

Issue: https://gitee.com/openharmony/third_party_llvm-project/issues/I8O7Q9
Signed-off-by: Nikolai Kholiavin <kholiavin.nikolai@huawei-partners.com>