[IR] Check for implicit 0 from Min module flag behavior in requirements
This fixes combination of Min IR module flag behavior and a module flag
requirement to work in the case where the flag is missing in one of the
modules. Previously this caused the requirement check to erroneously
succeed in IRLinker, but verifier would catch the problem and crash.

Issue: https://gitee.com/openharmony/third_party_llvm-project/issues/I9LG51

Signed-off-by: Nikolai Kholiavin <kholiavin.nikolai@huawei-partners.com>