[LLDB]Optimize the logic for "process" in linux to get loaded module
Description: check both th filename and directory when matching module in NativeProcessLinux::GetLoadedModuleFileSpec.
For example, if we loaded a module named "/A/liba.so" when attaching, and then "/B/liba.so" is dlopened during runtime. With this pr, lldb will not match "/A/liba.so" to "/B/liba.so".

Issue: https://gitee.com/openharmony/third_party_llvm-project/issues/I9RYMX
Test: dlopen a lib with the lib with the same name(but different dir) which loaded before attached. The new lib will not be binded to the old module.

Signed-off-by: liujialiang <liujialiang10@huawei.com>
Change-Id: Id15ab47e6540918759c391bb32fff3d71f79c067