[LLDB]Add a new property "target.modules-search-paths"
When debugging remotely, we always need to remap paths of modules in remote to local paths. LLDB has given command "target modules search-paths add" but this command can only be used after target created. If we want to attach to a process, this command can not be used before attaching.
So we add a property "target.modules-search-paths" to set the modules searching paths when initializing Target. User can set this property before command "attach" or "target create".
This property consists of an array of pairs, the first element is a path prefix, and the second is its replacement. The syntax is `prefix1 replacement1 prefix2 replacement2...`.  The pairs are checked in order, the first prefix that matches is used, and that prefix is substituted with the replacement.

Issue：https://gitee.com/openharmony/third_party_llvm-project/issues/IACS4H

Signed-off-by: liujialiang <liujialiang10@huawei.com>
Change-Id: I668a9a6f78df677852f4bf1236fa89122cdf21e1