NoPac: Fix Member function pointers
We need to mark virtual function member pointers also when
we do not signed the virtual function member pointer due to
NoPac.
We do not emit the thunk function for accessing the virtual
member function when the class is marked a NoPac.

origin author: Carlos Chinea <carlos.chinea.perez@huawei.com>
Signed-off-by: baojingjing <baojingjing@huawei.com>