[IRMover] Fix symver Copy @@
When FunctionImport uses IRMover to copy a function from other thinLTO
modules, it will copy the related .symver asm. However, like ".symver
foo, foo@@foo_1.0" sets the default version, requiring the corresponding
function to be defined. I believe that as a module using this function,
it only need to know the version corresponding to the symbol name it
uses, so the copied asm should be changed to ".symver foo, foo@foo_1.0".
To solve this problem, the code in IRMover needs to be modified.

Signed-off-by: Eric <chendewen3@huawei.com>